requirements.txt
astroid==2.2.5
boto==2.49.0
boto3==1.9.154
botocore==1.12.154
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
django-stubs==0.12.1
djangorestframework==3.8.2
djangorestframework-stubs==0.4.0
docutils==0.14
idna==2.7
isort==4.3.20
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
lazy-object-proxy==1.4.1
MarkupSafe==1.0
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.3.13
openapi-codec==1.3.2
Pillow==6.0.0
pylint==2.3.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.6.0
wrapt==1.11.1
